+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+* Add allow_update_secret parameter for instance secret updating
+
 # 3.18.1
 
 * Cleanup of AppDeployPipeline construct